feat(chat): add useChatUploadFiles hook, fix text content & minor refactor
