+++ apps/chat/src/components/Chat/ChatInput/ChatInput.tsx
@@ -8,10 +8,8 @@ import {
 
 import classNames from 'classnames';
 
+import { useChatUploadFiles } from '@/src/hooks/useChatUploadFiles';
 import { useFilePaste } from '@/src/hooks/useFilePaste';
-import { useUploadFilesHandler } from '@/src/hooks/useUploadFilesHandler';
-
-import { getQuickAttachmentsSavingPath } from '@/src/utils/app/conversation';
 
 import { ConversationsSelectors } from '@/src/store/conversations/conversations.reducers';
 import { useAppSelector } from '@/src/store/hooks';
@@ -55,23 +53,13 @@ export const ChatInput = Inversify.register(
     const messageIsStreaming = useAppSelector(
       ConversationsSelectors.selectIsConversationsStreaming,
     );
-    const availableAttachmentsTypes = useAppSelector(
-      ConversationsSelectors.selectAvailableAttachmentsTypes,
-    );
-    const maximumAttachmentsAmount = useAppSelector(
-      ConversationsSelectors.selectMaximumAttachmentsAmount,
-    );
     const canAttachFiles = useAppSelector(
       ConversationsSelectors.selectCanAttachFile,
     );
 
     const inputRef = useRef<HTMLDivElement | null>(null);
 
-    const handleUploadFiles = useUploadFilesHandler(
-      getQuickAttachmentsSavingPath(),
-      maximumAttachmentsAmount,
-      availableAttachmentsTypes,
-    );
+    const handleUploadFiles = useChatUploadFiles();
 
     const handlePaste = useCallback(
       (files: File[]) => {
